fixed textures freeing

+++ libft/ft_putnbr_fd.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(long int n, int fd)
 {
 	int		sign;
